Move ElasticaProductListener from CoreBundle to SearchBundle
