+++ src/Sylius/Bundle/SearchBundle/DependencyInjection/SyliusSearchExtension.php
@@ -11,15 +11,22 @@
 
 namespace Sylius\Bundle\SearchBundle\DependencyInjection;
 
+use FOS\ElasticaBundle\DependencyInjection\Configuration as FosElasticaConfiguration;
 use Sylius\Bundle\ResourceBundle\DependencyInjection\Extension\AbstractResourceExtension;
+use Sylius\Bundle\SearchBundle\DependencyInjection\Configuration as SyliusSearchConfiguration;
+use Sylius\Bundle\SearchBundle\Listener\ElasticaProductListener;
+use Symfony\Component\Config\Definition\Processor;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
+use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
+use Symfony\Component\DependencyInjection\Reference;
 
 /**
  * @author Argyrios Gounaris <agounaris@gmail.com>
  */
-class SyliusSearchExtension extends AbstractResourceExtension
+class SyliusSearchExtension extends AbstractResourceExtension implements PrependExtensionInterface
 {
     /**
      * {@inheritdoc}
@@ -53,5 +60,49 @@ class SyliusSearchExtension extends AbstractResourceExtension
         $container->setParameter('sylius_search.pre_search_filter.taxon', $config['filters']['pre_search_filter']['taxon']);
 
         $container->setParameter('sylius_search.custom.accessor.class', $config['custom_accessor']);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function prepend(ContainerBuilder $container)
+    {
+        $this->prependElasticaProductListener($container);
+    }
+
+    /**
+     * @param ContainerBuilder $container
+     */
+    private function prependElasticaProductListener(ContainerBuilder $container)
+    {
+        if (!$container->hasExtension('fos_elastica') || !$container->hasExtension('sylius_search')) {
+            return;
+        }
+
+        $configuration = new SyliusSearchConfiguration();
+        $processor = new Processor();
+        $syliusSearchConfig = $processor->processConfiguration($configuration, $container->getExtensionConfig('sylius_search'));
+        $engine = $syliusSearchConfig['engine'];
+
+        if ($engine === 'elasticsearch') {
+            $tags = ['doctrine.event_listener' => [
+                ['name' => 'doctrine.event_listener', 'event' => 'postPersist'],
+                ['name' => 'doctrine.event_listener', 'event' => 'postUpdate'],
+                ['name' => 'doctrine.event_listener', 'event' => 'postRemove'],
+                ['name' => 'doctrine.event_listener', 'event' => 'postFlush'],
+            ]];
+
+            $configuration = new FosElasticaConfiguration(false);
+            $processor = new Processor();
+            $elasticaConfig = $processor->processConfiguration($configuration, $container->getExtensionConfig('fos_elastica'));
+
+            foreach ($elasticaConfig['indexes'] as $index => $config) {
+                $elasticaProductListenerDefinition = new Definition(ElasticaProductListener::class);
+                $elasticaProductListenerDefinition->addArgument(new Reference('fos_elastica.object_persister.' . $index . '.product'));
+                $elasticaProductListenerDefinition->setTags($tags);
+
+                $container->setDefinition('sylius_product.listener.index_' . $index . '.product_update', $elasticaProductListenerDefinition);
+            }
+        }
     }
 }
